PHPCS: Fix most styling issues related to control structures
This enables the following sniffs:
 - Generic.ControlStructures.InlineControlStructure.NotAllowed
 - MediaWiki.ControlStructures.IfElseStructure.SpaceBeforeElse

NOTES:
This does *not* yet enable the last and following sniff:
MediaWiki.ControlStructures.AssignmentInControlStructures.AssignmentInControlStructures

This detects an issue where the $row variable is being assigned
the value of $dbr->fetchObject() in CosmosTemplate::getMostWantedPages().
This static function generally just needs to be refactored entirely,
and would be easier to fix everything in a separate, follow-up patch.

There are many nested control structures. This needs to be fixed
in a future follow-up patch.

Bug: T265108
Change-Id: I9b9515b18205f3f9dd5ac6a6eb761f223d920cbf