PHPCS: Fix most styling issues related to generic sniffs
This enables the following sniffs:
 - Generic.Arrays.DisallowLongArraySyntax.Found
 - Generic.Functions.FunctionCallArgumentSpacing.SpaceBeforeComma
 - Generic.WhiteSpace.DisallowSpaceIndent.SpacesUsed

There are enough errors generated from Generic.Files.LineLength.TooLong
that it warrants it's own follow-up patch. It is also separate since it
is not an issue that can be auto-fixed with PHPCS. It could either be
fixed by itself, with the MediaWiki whitespace sniffs, Squiz whitespace
sniffs, or one giant patch for all whitespace sniffs/whitespace issues.

Bug: T265108
Change-Id: Id3bc29c9f1965f2df7942a1af3481f5b82aa0645