CosmosTemplate: Don't pass string parameter by reference
Originally, all of these functions declared their method signature
to have it so that a variable is passed by reference. This is very
fragile, and instead, each method should be returning a string
instead of modifying an existing variable. This allows each method
to exist on their own, so they are not tightly coupled to each
other. Each method is easier to unit test on their own.

Bug: T265232
Change-Id: Ib72a92eb4aede2ec591c506d8406053a24608386