Replace most calls to the global wfMessage() function
This inject a MessageLocalizer into CosmosNavigation and CosmosToolbar.
This helps replaces calls to wfMessage() in both classes. Currently,
these do *not* replace calls to:
 - CosmosNavigation::parseItem(),
 - CosmosToolbar::parseItem()

This is because the parseItem() is a static function, and thus
can't access a MessageLocalizer. These functions are also both
unused, so it doesn't currently make sense to replace these when
they're most likely going to be removed in a follow-up patch.

Not all dependencies are injected (e.g TitleFactory, HookRunner, etc.)
That would be done in a follow-up patch. Ideally though, both of these
classes should be refactored to deduplicate logic and make it easier
to maintain, read, etc.

This does *not* replace the wfMessage() call in the AdminLinks hook
listener, because:
  - That hook should probably be providing access to the MessageLocalizer
    interface instead
  - Creating an instance of RequestContext is more expensive than calling
    wfMessage()

Bug: T265195
Bug: T265110
Change-Id: I5c18f53d3caffddb0a76685128cd5cfd35e861b4