Make user skin preference for Cosmos consistent with conventions
The current state had it so that the Cosmos preference had its
own tab in Special:Preferences. However, this is not the convention.

The convention is to have all skin preferences under the "Appearance"
tab, which is where the user can choose their skin, where there will
also be a skin preferences subsection.

This patch makes it so that it's consistent with the convention, and
so that all user preferences can be expected to only stay in one spot.

An i18n message becomes unused in this patch, and is removed from the
en.json and qqq.json, as Translatewiki.net will handle the rest of
the i18n JSON files automatically.

Change-Id: I6f2737a872a869fd788a803dbb5caa485b323393