Simplify check for seeing if the Title instance is main page
In the OutputPageBodyAttributes hook listener, the source code currently
instantiates a new Title object via Title::newMainPage(). We can save
performance by using an existing method in the Title object to verify
that it's a title object instead, via Title->isMainPage().

NOTES:
This code is actually redundant, as MediaWiki core will still auto-add the
page name as a class to the body element. The skin class will add these:
 - page-{page name}
 - rootpage-{root page name}

This needs to be fixed in a follow-up patch of Cosmos, as it also requires
adjusting a LESS stylesheet.

Change-Id: If9afee34d5ff6e3703f0b18def0519869f8798d8