Use guard clause for readability in CosmosTemplate->buildFooterIcons()
Some minor preparation work for the eventual work of T264908.

Guard clauses help reduce the amount of indentation and improve
the readability of code. Using a guard clause does not change
the behavior of the code, it simply creates an early return.
The main logic occurs after the expression, if the expression
fails.

Bug: T264908
Change-Id: I21c788919ce642efcfa61a701cc343e7c26893bb